+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/rejectorder/RejectedConsentOrderAboutToSubmitHandler.java
@@ -1,34 +1,41 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.handler;
+package uk.gov.hmcts.reform.finrem.caseorchestration.handler.rejectorder;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.finrem.caseorchestration.ccd.callback.CallbackType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.controllers.GenericAboutToStartOrSubmitCallbackResponse;
+import uk.gov.hmcts.reform.finrem.caseorchestration.handler.CallbackHandlerLogger;
+import uk.gov.hmcts.reform.finrem.caseorchestration.handler.FinremAboutToSubmitCallbackHandler;
+import uk.gov.hmcts.reform.finrem.caseorchestration.handler.FinremCallbackRequest;
 import uk.gov.hmcts.reform.finrem.caseorchestration.mapper.FinremCaseDetailsMapper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseData;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseDetails;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.RefusalOrderDocumentService;
+import uk.gov.hmcts.reform.finrem.caseorchestration.service.documentcatergory.RefusedConsentOrderDocumentCategoriser;
 
 @Slf4j
 @Service
-public class RejectedConsentOrderMidHandler extends FinremCallbackHandler {
+public class RejectedConsentOrderAboutToSubmitHandler extends FinremAboutToSubmitCallbackHandler {
 
     private final RefusalOrderDocumentService service;
+    private final RefusedConsentOrderDocumentCategoriser categoriser;
 
     @Autowired
-    public RejectedConsentOrderMidHandler(FinremCaseDetailsMapper mapper,
-                                          RefusalOrderDocumentService service) {
+    public RejectedConsentOrderAboutToSubmitHandler(FinremCaseDetailsMapper mapper,
+                                                    RefusalOrderDocumentService service,
+                                                    RefusedConsentOrderDocumentCategoriser categoriser) {
         super(mapper);
         this.service = service;
+        this.categoriser = categoriser;
     }
 
     @Override
     public boolean canHandle(final CallbackType callbackType, final CaseType caseType,
                              final EventType eventType) {
-        return CallbackType.MID_EVENT.equals(callbackType)
+        return CallbackType.ABOUT_TO_SUBMIT.equals(callbackType)
             && CaseType.CONSENTED.equals(caseType)
             && EventType.REJECT_ORDER.equals(eventType);
     }
@@ -36,23 +43,12 @@ public class RejectedConsentOrderMidHandler extends FinremCallbackHandler {
     @Override
     public GenericAboutToStartOrSubmitCallbackResponse<FinremCaseData> handle(FinremCallbackRequest callbackRequest,
                                                                               String userAuthorisation) {
+        log.info(CallbackHandlerLogger.aboutToSubmit(callbackRequest));
         FinremCaseDetails caseDetails = callbackRequest.getCaseDetails();
-        log.info("Received request for '{}' event '{}' for Case ID: {}",CallbackType.MID_EVENT,
-            EventType.REJECT_ORDER, caseDetails.getId());
 
-        FinremCaseData caseData = service.previewConsentOrderNotApproved(userAuthorisation, caseDetails);
-        clearContestedFields(caseData);
-        return GenericAboutToStartOrSubmitCallbackResponse.<FinremCaseData>builder()
-            .data(caseData)
-            .build();
-    }
+        FinremCaseData caseData = service.processConsentOrderNotApproved(caseDetails, userAuthorisation);
+        categoriser.categorise(caseData);
 
-    private void clearContestedFields(FinremCaseData caseData) {
-        if (caseData != null) {
-            caseData.setIntervenerOne(null);
-            caseData.setIntervenerTwo(null);
-            caseData.setIntervenerThree(null);
-            caseData.setIntervenerFour(null);
-        }
+        return response(caseData);
     }
 }
