+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/handler/rejectorder/RejectedConsentOrderAboutToSartHandlerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.handler;
+package uk.gov.hmcts.reform.finrem.caseorchestration.handler.rejectorder;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -6,6 +6,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.reform.finrem.caseorchestration.ccd.callback.CallbackType;
+import uk.gov.hmcts.reform.finrem.caseorchestration.handler.FinremCallbackRequest;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.EventType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseData;
@@ -15,6 +16,7 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.service.RefusalOrderDocument
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.verify;
+import static uk.gov.hmcts.reform.finrem.caseorchestration.TestConstants.AUTH_TOKEN;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CaseType.CONTESTED;
 
 @ExtendWith(MockitoExtension.class)
@@ -24,8 +26,6 @@ class RejectedConsentOrderAboutToSartHandlerTest {
     private RejectedConsentOrderAboutToSartHandler handler;
     @Mock
     private RefusalOrderDocumentService refusalOrderDocumentService;
-
-    private static final String AUTH_TOKEN = "Token-:";
 
     @Test
     void given_case_whenEventRejectedOrder_thenCanHandle() {
